% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convcheck.R
\name{convcheck}
\alias{convcheck}
\title{Assess convergence of Bayesian Bootstrap from an fwb object object}
\usage{
convcheck(fwbobj, tix = "random")
}
\arguments{
\item{fwbobj}{Result of running fwb with do_one_bbdwols function}

\item{tix}{index of blip parameters to plot. Either numeric vector,
"random" which will randomly choose 1/4 of parameters to check, or
"all" which will plot all parameters}
}
\value{
Empty
}
\description{
Assess convergence of Bayesian Bootstrap from an fwb object object
}
