% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_one_bbdwols.R
\name{do_one_bbdwols}
\alias{do_one_bbdwols}
\title{Execute 1 Bayesian Bootstrap iteration of DWOLS given data and Dirichlet weights w}
\usage{
do_one_bbdwols(
  data,
  w,
  outcome.mod,
  trt.mod,
  maxit = 500,
  missingoutcome = FALSE,
  outmiss.mod = NULL,
  scale_vars = TRUE,
  trim_weights = 1,
  mode = "parameter"
)
}
\arguments{
\item{data}{Data.frame of study data}

\item{w}{vector of weights, one for every row in data, drawn from a Dirichlet (1,...,1) distribution. Supplied by fwb}

\item{outcome.mod}{A formula for the outcome model using variable names in data. Outcome name extracted from LHS}

\item{trt.mod}{A formula for modelling treatment assignment using variable names in data. Treatment assignment extracted from LHS}

\item{maxit}{Maximum number of iterations to allow for fitting outcome model using nnet}

\item{missingoutcome}{Logical for whether weights that consider the missing outcomes should be used. Otherwise, just uses standard DWOLS IPT weights.}

\item{outmiss.mod}{A formula for modelling missing outcomes using variables in data. Missing outcomes indicated with NA in response column}

\item{scale_vars}{Logical for if variables should be scaled to approximately between 0,1 when fitting treatment assignment model to improve convergence of nnet}

\item{trim_weights}{Percentile at which to trim weights. Default is 1 (no trimming).}

\item{mode}{Character string indicating what to return. Default is "parameter" which returns blip parameter posterior samples.
Other options useful for debugging or assessing suitability of models are "iptw" which returns inverse probability of treatment, "ipmw" which returns inverse probability of being missing,
and "allweights" which returns the weights used in finding the blip parameter estimates.}
}
\description{
Do 1 DWOLS with Bayesian Bootstrap weights. Designed for use with the fwb R package.
}
\examples{

if(require("fwb")) {
# Create example data
example_dat <- data.frame(X1 = runif(200),
                          X2 = rnorm(200, 8),
                          trt = sample(c("1", "2", "3"), 200, replace = TRUE))
example_dat$outcome <- 5 + 0.1*example_dat$X1 + 0.2*example_dat$X2 +
                         as.numeric(example_dat$trt) +
                         as.numeric(example_dat$trt)*example_dat$X1

# BBDWOLS
result <- fwb(data = example_dat,
              statistic = do_one_bbdwols,
              outcome.mod = outcome ~ X2 + trt*X1,
              trt.mod = trt ~ 1,
              R = 199,
              missingoutcome = FALSE,
              mode = "parameter")
convcheck(result, tix = c(1,2))

}

}
