% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.bbdwols.R
\name{one.bbdwols}
\alias{one.bbdwols}
\title{Function that executes 1 iteration of bbdwols}
\usage{
one.bbdwols(
  data,
  w,
  outcome.mod,
  trt.mod,
  maxit = 500,
  missingoutcome = FALSE,
  outmiss.mod = NULL,
  scale_vars,
  mode = "parameter"
)
}
\description{
Function that executes 1 iteration of bbdwols
}
