% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbdwols.R
\name{bbdwols}
\alias{bbdwols}
\title{Bayesian Bootstrap dWOLS}
\usage{
bbdwols(outcome.mod, trt.mod, dat, L, trt.name, maxit = 500, diag = TRUE, ...)
}
\arguments{
\item{outcome.mod}{A formula for the outcome model}

\item{trt.mod}{A Formula for the treatment model}

\item{dat}{the data.frame with all data, must have variables for outcome and treatment model}

\item{L}{The number of bootstrap samples to take}

\item{trt.name}{A string giving the name of the outcome in dat}

\item{diag}{Logical indicating if diagnostic plots to assess convergence of bootstrap algorithm should be made, default TRUE}
}
\value{
A data.frame with columns corresponding to parameters in the outcome model and each row is an iteration of the bbdwols
}
\description{
Bayesian Bootstrap dWOLS
}
