% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbdwols.R
\name{bbdwols}
\alias{bbdwols}
\title{Bayesian Bootstrap dWOLS}
\usage{
bbdwols(
  outcome.mod,
  trt.mod,
  dat,
  L,
  missingoutcome,
  outmiss.mod = NULL,
  maxit = 500,
  diag = TRUE,
  full_posterior = FALSE,
  ...
)
}
\arguments{
\item{outcome.mod}{A formula for the outcome model}

\item{trt.mod}{A Formula for the treatment model}

\item{dat}{the data.frame with all data, must have variables for outcome and treatment model}

\item{L}{The number of bootstrap samples to take}

\item{missingoutcome}{Logical indicating if probability of missingness weighting should be used to account for missing outcome data. If FALSE, observations with missing outcomes are removed for the analysis.}

\item{outmiss.mod}{A formula describing the explanatory variables to include in the outcome missingness model, of the format ~ (explanatory variables)}

\item{diag}{Logical indicating if diagnostic plots to assess convergence of bootstrap algorithm should be made, default TRUE}

\item{full_posterior}{Logical indicating if full posterior samples should be returned, or just point estimates. Default is FALSE}
}
\value{
A data.frame with columns corresponding to parameters in the outcome model and each row is an iteration of the bbdwols
}
\description{
Bayesian Bootstrap dWOLS
}
