% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convcheck.R
\name{convcheck}
\alias{convcheck}
\title{Assess convergence of Bayesian Bootstrap for a bbdwols object}
\usage{
convcheck(obj, ...)
}
\arguments{
\item{obj}{A bbdwols object}

\item{...}{additional arguments to \code{par}}
}
\value{
Empty
}
\description{
Assess convergence of Bayesian Bootstrap for a bbdwols object
}
